packages/session-memory/pyproject.toml
# SPDX-License-Identifier: Apache-2.0
# Copyright 2026 CervellaSwarm Contributors

[build-system]
requires = ["hatchling>=1.26"]
build-backend = "hatchling.build"

[project]
name = "cervellaswarm-session-memory"
version = "0.1.0"
description = "Git-native, human-readable session memory for AI agents. Every decision is a commit, every state is Markdown."
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE", "NOTICE"]
requires-python = ">=3.10"
authors = [{ name = "CervellaSwarm Contributors" }]
keywords = [
    "session-memory",
    "ai-agents",
    "multi-agent",
    "git-native",
    "markdown",
    "audit-trail",
    "session-continuity",
    "claude-code",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = ["pyyaml>=6.0"]

[project.scripts]
cervella-session = "cervellaswarm_session_memory.cli:main"
cervella-session-init = "cervellaswarm_session_memory.cli:main_init"
cervella-session-check = "cervellaswarm_session_memory.cli:main_check"
cervella-session-audit = "cervellaswarm_session_memory.cli:main_audit"
cervella-session-sync = "cervellaswarm_session_memory.cli:main_sync"

[project.urls]
Homepage = "https://github.com/rafapra3008/cervellaswarm"
Repository = "https://github.com/rafapra3008/cervellaswarm"
Documentation = "https://github.com/rafapra3008/cervellaswarm/tree/main/packages/session-memory"
"Bug Tracker" = "https://github.com/rafapra3008/cervellaswarm/issues"

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.9.0"]
test = ["pytest>=8.0", "pytest-cov>=5.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/cervellaswarm_session_memory"]

[tool.hatch.build.targets.sdist]
include = ["src/", "tests/", "LICENSE", "NOTICE", "README.md", "pyproject.toml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = ["--tb=short", "-q"]

[tool.coverage.run]
source = ["cervellaswarm_session_memory"]
omit = ["*/tests/*", "*/__main__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true
